Add inference rules for SEQSCAN
SEQSCAN data structures can now be infered such that
a valid plan is generated. Some parts of these rules
are pretty static and just use constants for some fields
but that is fine considering the current InAST.

In-depth postgres src analysis will be required to identify
which of these fields have to be accurate.