Add LIMIT operator
LIMIT can now be used as operator.
Validation of this node is incomplete.
Ideally it should be checked if result yields a values of
the type int8, but that's out of scope at this stage.
It would be possible for CONSTS, but we will have to revisit
this part of the code  as soon as we can build more complex
expressions (such as function calls, operators etc.)