Add FUNCEXPR inference rule
We can now use function calls.

Fix wrong data type representation of qual.
We had this as `Maybe Expr` but it has to be
`List Expr`, where the elements are implicitly
ANDed conditions. Exception to this is RESULT,
because we do `:resconstantqual` instead of `:qual`.

Validation of function calls is very weak!
No type checks are performed. Just the number
of arguments are checked.